+++ src/view/IDisplay.java
@@ -23,7 +23,7 @@ public interface IDisplay {
 
 	public IModel getModel();
 	
-	public int[] inputPopup(String[] message) ;
+	public double[] inputPopup(String[] message) ;
 	
 	public int getScale();
 }
